+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "stars-circles",
-  "version": "1.0.0",
+  "version": "2.2.0",
   "description": "",
   "main": "index.js",
   "scripts": {
