+++ src/App.css
@@ -17,23 +17,27 @@ html {
   --red-color-hint: rgb(173 107 71 / 30%);
   --black-color: rgb(0 0 0);
   --default-font-size: min(0.6cm, 5vmin);
-  --num-columns: 5;
   --square-diameter: calc(var(--width) / var(--num-columns));
 
   font-family: Arial, Helvetica, sans-serif;
   justify-items: center;
   font-size: var(--default-font-size);
+  height: 95vh;
+  width: var(--width);
+  justify-self: center;
+  white-space: pre-wrap;
+}
+
+.game {
   display: grid;
+
   grid-template-areas:
     "score"
     "board"
     "tokens"
     "controls";
   grid-template-rows: auto auto 1fr auto;
-  height: 95vh;
-  width: var(--width);
-  justify-self: center;
-  white-space: pre-wrap;
+
 }
 
 button {
@@ -49,6 +53,10 @@ button {
   color: black;
   margin: 0;
   padding: 1px 6px;
+}
+
+button:disabled {
+  opacity: 0.6;
 }
 
 .board {
@@ -154,22 +162,6 @@ button {
   background-image: url("./images/undo.svg");
 }
 
-.modal {
-  background-image: url("./images/background.svg");
-  position: fixed;
-  left: 0;
-  top: 0;
-  height: 100%;
-  width: 100%;
-  text-align: center;
-  overflow: scroll;
-  justify-items: center;
-  justify-content: space-evenly;
-  align-items: center;
-  display: grid;
-  grid-template-rows: 1fr auto;
-}
-
 #settings {
   display: flex;
   flex-direction: column;
@@ -189,7 +181,7 @@ button {
 }
 
 .setting-info {
-  font-size: calc(var(--default-font-size) * 0.5);
+  font-size: calc(var(--default-font-size) * 0.7);
 }
 
 #setting-buttons {
@@ -203,10 +195,14 @@ button {
   margin: 3vmin;
 }
 
-#rules-modal {
+.rules {
   display: flex;
   flex-direction: column;
+  text-align: center;
+  overflow: scroll;
   justify-content: space-between;
+  align-items: center;
+  grid-template-rows: 1fr auto;
 }
 
 .rules-text {
